requirements.txt
python-telegram-bot
discord.py
pywhatkit
python-dotenv
requests
transformers
torch